+++ src/base/SimpleCutMan.cpp
@@ -1,12 +1,12 @@
-// 
+//
 //     Minotaur -- It's only 1/2 bull
-// 
+//
 //     (C)opyright 2010 - 2025 The Minotaur Team.
-// 
+//
 
 /**
  * \file SimpleCutMan.cpp
- * \brief Implement the methods of SimpleCutMan class. 
+ * \brief Implement the methods of SimpleCutMan class.
  * \author Ashutosh Mahajan, IIT Bombay
  */
 
@@ -30,12 +30,12 @@ using namespace Minotaur;
 typedef std::list<CutPtr>::iterator CLIter;
 typedef std::vector<ConstraintPtr>::const_iterator ConstIter;
 
-const std::string SimpleCutMan::me_ = "SimpleCutMan: "; 
+const std::string SimpleCutMan::me_ = "SimpleCutMan: ";
 
 
 SimpleCutMan::SimpleCutMan()
-  : env_(EnvPtr()),   // NULL
-    p_(ProblemPtr()), // NULL
+  : env_(EnvPtr()),    // NULL
+    p_(ProblemPtr()),  // NULL
     violAbs_(1e-4),
     violRel_(1e-3)
 {
@@ -76,31 +76,31 @@ ConstraintPtr SimpleCutMan::addCut(ProblemPtr p, FunctionPtr f, double lb,
 
 void SimpleCutMan::addCuts(CutVectorIter cbeg, CutVectorIter cend)
 {
-  for (CutVectorIter it=cbeg; it!=cend; ++it) {
+  for (CutVectorIter it = cbeg; it != cend; ++it) {
     addCut(*it);
   }
 }
 
 
-UInt SimpleCutMan::getNumCuts() const
+size_t SimpleCutMan::getNumCuts() const
 {
   return enCuts_ + newCuts_.size();
 }
 
 
-UInt SimpleCutMan::getNumEnabledCuts() const
+size_t SimpleCutMan::getNumEnabledCuts() const
 {
   return enCuts_;
 }
 
 
-UInt SimpleCutMan::getNumDisabledCuts() const
+size_t SimpleCutMan::getNumDisabledCuts() const
 {
   return 0;
 }
 
 
-UInt SimpleCutMan::getNumNewCuts() const
+size_t SimpleCutMan::getNumNewCuts() const
 {
   return newCuts_.size();
 }
@@ -108,7 +108,7 @@ UInt SimpleCutMan::getNumNewCuts() const
 
 void SimpleCutMan::mvNewToPool_()
 {
-  for (CLIter it=newCuts_.begin(); it!=newCuts_.end(); ++it) {
+  for (CLIter it = newCuts_.begin(); it != newCuts_.end(); ++it) {
     pool_.push_back(*it);
   }
 }
@@ -119,7 +119,8 @@ void SimpleCutMan::postSolveUpdate(ConstSolutionPtr, EngineStatus)
 }
 
 
-void SimpleCutMan::separate(ProblemPtr p, ConstSolutionPtr sol, bool *, UInt *)
+void SimpleCutMan::separate(ProblemPtr p, ConstSolutionPtr sol, bool *,
+                            size_t *)
 {
   const double *x = sol->getPrimal();
   CutPtr cut;
@@ -127,27 +128,27 @@ void SimpleCutMan::separate(ProblemPtr p, ConstSolutionPtr sol, bool *, UInt *)
   int err;
 
   mvNewToPool_();
- 
-  for (CLIter it=pool_.begin(); it!=pool_.end();) {
+
+  for (CLIter it = pool_.begin(); it != pool_.end();) {
     cut = *it;
 #if SPEW
     cut->write(logger_->msgStream(LogInfo));
 #endif
     err = 0;
     act = cut->eval(x, &err);
-    if (err!=0) {
-      logger_->msgStream(LogInfo) << me_ << "Error evaluating activity of cut. "
-                                  << "Not adding to relaxation. Cut is: "
-                                  << std::endl;
+    if (err != 0) {
+      logger_->msgStream(LogInfo)
+          << me_ << "Error evaluating activity of cut. "
+          << "Not adding to relaxation. Cut is: " << std::endl;
       ++it;
       continue;
     }
-    viol = std::max(cut->getLb()-act, act-cut->getUb());
-    if (viol > violAbs_ + violRel_*fabs(act)) {
+    viol = std::max(cut->getLb() - act, act - cut->getUb());
+    if (viol > violAbs_ + violRel_ * fabs(act)) {
 #if SPEW
-      logger_->msgStream(LogInfo) << me_ << "Solution violates cut, "
-                                  << "adding cut to relaxation. Cut is: "
-                                  << std::endl;
+      logger_->msgStream(LogInfo)
+          << me_ << "Solution violates cut, "
+          << "adding cut to relaxation. Cut is: " << std::endl;
 #endif
       ++enCuts_;
       cut->applyToProblem(p);
@@ -173,5 +174,3 @@ void SimpleCutMan::writeStats(std::ostream &out) const
 {
   out << me_ << "No stats availale" << std::endl;
 }
-
-
